♻️: refactor 모두의 답변 조회에 적합한 네이밍으로 클래스,메서드 네이밍 변경 #32
