+++ plu-api/src/main/kotlin/com/th/plu/api/controller/answer/AnswerController.kt
@@ -3,7 +3,7 @@ package com.th.plu.api.controller.answer
 import com.th.plu.api.config.interceptor.Auth
 import com.th.plu.api.config.resolver.MemberId
 import com.th.plu.api.controller.answer.dto.response.AnswerInfoResponse
-import com.th.plu.api.controller.answer.dto.response.AnswerRetrievePageNationResponses
+import com.th.plu.api.controller.answer.dto.response.EveryAnswerRetrievePageNationResponses
 import com.th.plu.api.service.answer.AnswerService
 import com.th.plu.common.dto.response.ApiResponse
 import io.swagger.v3.oas.annotations.Operation
@@ -45,7 +45,7 @@ class AnswerController(
     fun pageAnswers(
             @RequestParam lastAnswerId: Long,
             @RequestParam pageSize: Long,
-    ): ApiResponse<AnswerRetrievePageNationResponses> {
-        return ApiResponse.success(answerService.retrieveTodayAnswersWithCursor(lastAnswerId, pageSize))
+    ): ApiResponse<EveryAnswerRetrievePageNationResponses> {
+        return ApiResponse.success(answerService.retrieveEveryAnswersWithCursor(lastAnswerId, pageSize))
     }
 }
