--- plu-api/src/main/kotlin/com/th/plu/api/controller/answer/dto/response/AnswerRetrievePageNationResponse.kt
@@ -1,4 +1,0 @@
-package com.th.plu.api.controller.answer.dto.response
-
-data class AnswerRetrievePageNationResponse(val answerId: Long, val likeCount: Long, val answerContent: String) {
-}
