+++ plu-api/src/main/kotlin/com/th/plu/api/controller/answer/dto/response/EveryAnswerRetrievePageNationResponse.kt
@@ -1,0 +1,4 @@
+package com.th.plu.api.controller.answer.dto.response
+
+data class EveryAnswerRetrievePageNationResponse(val answerId: Long, val likeCount: Long, val answerContent: String) {
+}
